major update, fixed lots of force close issues,
tablets 100% supported now, preparing to include AdMob support

+++ MyCalcFinance/src/com/mwerner/mycalc/finance/F_List.java
@@ -26,8 +26,7 @@ public class F_List extends ListFragment {
 		// check to see if there are the frame layouts (i.e. is the device in
 		// landscape?
 		View detailsFrame = getActivity().findViewById(R.id.content1);
-		mDualPane = detailsFrame != null
-				&& detailsFrame.getVisibility() == View.VISIBLE;
+		mDualPane = detailsFrame != null && detailsFrame.getVisibility() == View.VISIBLE;
 
 	}
 
@@ -38,7 +37,7 @@ public class F_List extends ListFragment {
 		 * here the long ID is converted into a int for switch statements here
 		 * and in the DetailsActivity
 		 */
-		Integer id2 = (int) (long) id;
+		Integer id2 = (int) id;
 
 		if (mDualPane) {
 			/*
@@ -53,8 +52,7 @@ public class F_List extends ListFragment {
 			case 0:
 				Fragment frag0 = new F_PV();
 				Fragment frag01 = new F_PV_Info();
-				FragmentTransaction ft = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft = getActivity().getSupportFragmentManager().beginTransaction();
 				ft.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft.replace(R.id.content1, frag0);
 				ft.replace(R.id.content2, frag01);
@@ -63,8 +61,7 @@ public class F_List extends ListFragment {
 			case 1:
 				Fragment frag1 = new F_PVAnn();
 				Fragment frag11 = new F_PVAnn_Info();
-				FragmentTransaction ft1 = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft1 = getActivity().getSupportFragmentManager().beginTransaction();
 				ft1.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft1.replace(R.id.content1, frag1);
 				ft1.replace(R.id.content2, frag11);
@@ -73,8 +70,7 @@ public class F_List extends ListFragment {
 			case 2:
 				Fragment frag2 = new F_PVAnnDue();
 				Fragment frag21 = new F_PVAnnDue_Info();
-				FragmentTransaction ft2 = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft2 = getActivity().getSupportFragmentManager().beginTransaction();
 				ft2.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft2.replace(R.id.content1, frag2);
 				ft2.replace(R.id.content2, frag21);
@@ -83,8 +79,7 @@ public class F_List extends ListFragment {
 			case 3:
 				Fragment frag3 = new F_FVAnn();
 				Fragment frag31 = new F_FVAnn_Info();
-				FragmentTransaction ft3 = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft3 = getActivity().getSupportFragmentManager().beginTransaction();
 				ft3.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft3.replace(R.id.content1, frag3);
 				ft3.replace(R.id.content2, frag31);
@@ -93,8 +88,7 @@ public class F_List extends ListFragment {
 			case 4:
 				Fragment frag4 = new F_FVAnnDue();
 				Fragment frag41 = new F_FVAnnDue_Info();
-				FragmentTransaction ft4 = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft4 = getActivity().getSupportFragmentManager().beginTransaction();
 				ft4.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft4.replace(R.id.content1, frag4);
 				ft4.replace(R.id.content2, frag41);
@@ -103,8 +97,7 @@ public class F_List extends ListFragment {
 			case 5:
 				Fragment frag5 = new F_Bonds();
 				Fragment frag51 = new F_Bonds_Info();
-				FragmentTransaction ft5 = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft5 = getActivity().getSupportFragmentManager().beginTransaction();
 				ft5.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft5.replace(R.id.content1, frag5);
 				ft5.replace(R.id.content2, frag51);
@@ -113,8 +106,7 @@ public class F_List extends ListFragment {
 			case 6:
 				Fragment frag6 = new F_Growth();
 				Fragment frag61 = new F_Growth_Info();
-				FragmentTransaction ft6 = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft6 = getActivity().getSupportFragmentManager().beginTransaction();
 				ft6.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft6.replace(R.id.content1, frag6);
 				ft6.replace(R.id.content2, frag61);
@@ -123,8 +115,7 @@ public class F_List extends ListFragment {
 			case 7:
 				Fragment frag7 = new F_NPV();
 				Fragment frag71 = new F_NPV_Info();
-				FragmentTransaction ft7 = getActivity()
-						.getSupportFragmentManager().beginTransaction();
+				FragmentTransaction ft7 = getActivity().getSupportFragmentManager().beginTransaction();
 				ft7.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
 				ft7.replace(R.id.content1, frag7);
 				ft7.replace(R.id.content2, frag71);
